refactor: Add the prefix _ to the filenames of non-public modules
