+++ docs/api/astrodynx.twobody.rst
@@ -1,40 +1,10 @@
 ``astrodynx.twobody``
 ======================
 
-``twobody``
------------
-
 .. automodule:: astrodynx.twobody
     :members:
     :undoc-members:
     :show-inheritance:
-
-``twobody.orb_integrals``
--------------------------
-
-.. automodule:: astrodynx.twobody.orb_integrals
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-``twobody.kep_equ``
--------------------
-
-.. automodule:: astrodynx.twobody.kep_equ
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-``twobody.ivp``
----------------
-
-.. automodule:: astrodynx.twobody.ivp
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-``twobody.uniformulas``
------------------------
 
 .. automodule:: astrodynx.twobody.uniformulas
     :members:
